style: resolve cargo clippy -D warnings on master
Fixes 18 clippy errors that surfaced once the lint job started gating
master. Mostly mechanical, applied via cargo clippy --fix:

- io_other_error in wipe/selfdelete.rs
- single_char_add_str in ai/builder.rs, github/{mod,types}.rs
- derivable_impls for EnvironmentType, RiskTolerance, TeamSize in ai/types.rs
- collapsible_if in retention/mod.rs
- manual_contains in accounts/validation.rs
- unnecessary_cast in eventlog/storage.rs
- unnecessary_lazy_evaluations in reports/exporter.rs
- nonminimal_bool in reports/settings.rs
- needless_question_mark in reports/pdf.rs
- unused_assignments in reports/pdf.rs (idx scoped to loop body)
- unnecessary_sort_by in reports/aggregator.rs (sort_by_key + Reverse)

No semantic behaviour changes. cargo fmt + cargo clippy -D warnings +
cargo test --lib all clean locally.

+++ src-tauri/src/eventlog/storage.rs
@@ -226,7 +226,7 @@ pub fn purge_older_than(cutoff: DateTime<Utc>) -> Result<usize, EventLogError> {
         "DELETE FROM event_log WHERE occurred_at < ?1",
         params![cutoff.to_rfc3339()],
     )?;
-    Ok(affected as usize)
+    Ok(affected)
 }
 
 /// Total row count. Used by the QA contract's responsiveness check.
